CRBUG
Former-commit-id: 02b0430c1fc5c95f8f12d8ea0828e67ad7a1bb6a

+++ ELA/found/SuperPoint/view/HMWinputVotesPopupWindowView.h
@@ -18,7 +18,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface HMWinputVotesPopupWindowView : UIView
 @property(nonatomic,weak)id<VotesPopupViewDelegate> delegate;
 @property (weak, nonatomic) IBOutlet UILabel *accountBalanceLab;
-@property(nonatomic,assign)NSInteger votes;
+@property(nonatomic,assign)double votes;
 
 
 @end
